backend/pyproject.toml
[tool.poetry]
name = "divvy-backend"
version = "0.1.0"
description = ""
authors = ["Jaric Thorning <jaric.thorning@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.1"
python-decouple = "^3.8"
psycopg2 = "^2.9.9"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
